COMP: Fix -Wsign-compare in vtkSubjectHierarchyItem::AddToTree
This commit fixes a warning originally introduced in ac3d96b99 (ENH: Make
Clone action clone a segment when a segment is selected in subject
hierarchy).

+++ Libs/MRML/Core/vtkMRMLSubjectHierarchyNode.cxx
@@ -339,7 +339,7 @@ vtkIdType vtkSubjectHierarchyItem::AddToTree(vtkSubjectHierarchyItem* parent, st
     {
     // Add under parent
     vtkSmartPointer<vtkSubjectHierarchyItem> childPointer(this);
-    if (positionUnderParent<0 || positionUnderParent >= this->Parent->Children.size())
+    if (positionUnderParent < 0 || positionUnderParent >= static_cast<int>(this->Parent->Children.size()))
       {
       this->Parent->Children.push_back(childPointer);
       }
